KafkaMessageConsumer gets its own feature
