+++ src/Harald.WebApi/Features/KafkaMessageConsumer/Infrastructure/KafkaConsumerFactory.cs
@@ -1,11 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using Confluent.Kafka;
 using Microsoft.Extensions.Configuration;
 
-namespace Harald.WebApi.Infrastructure.Messaging
+namespace Harald.WebApi.Features.KafkaMessageConsumer.Infrastructure
 {
     public class KafkaConsumerFactory
     {
